Resolve FMath function ambiguities
git-svn-id: file:///A/Documents/Unreal%20Projects/RogueSkySVNtoGit@14 12684d69-bc46-6143-8e79-893cf67d636b

Former-commit-id: 9eb4a813fa97e223af88f8e3e1c6aeeadb2c7d76

+++ Source/RogueSky/Private/Generation/Chunk.cpp
@@ -50,7 +50,7 @@ FVector Chunk::GetVertexLocation(int8 X, int8 Y, int8 Z, uint8 LODLevel) const {
     if (LODLevel == 0)
         return GetVertexLocation(X, Y, Z);
 
-    int lodUnitSize = FMath::Exp2(LODLevel);
+    uint8 lodUnitSize = FMath::Exp2((float)LODLevel);
 
     FVector vectorSum = FVector::ZeroVector;
     int totalVectors = 0;
